pinup: Move to separate name space, use own option and clean up code.
